make weight a byte

+++ src/bookwrit.h
@@ -32,7 +32,7 @@ class BookWriter {
         ~BookWriter();
                 
         // add a move to the file.
-        void add(const hash_t hashCode, byte moveIndex, uint16_t weight,
+        void add(const hash_t hashCode, byte moveIndex, byte weight,
                  uint32_t win, uint32_t loss, uint32_t draw);
 
         // Write book contents out to the designated path. Returns 0
